requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-rerunfailures==10.2
python-dotenv==1.0.0
requests==2.28.2
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.6
wsproto==1.2.0
